Extract space/character iterator from renderer
