+++ src/alignment/center.rs
@@ -3,8 +3,7 @@ use crate::{
     alignment::{HorizontalTextAlignment, VerticalTextAlignment},
     parser::{Parser, Token},
     rendering::{
-        cursor::Cursor,
-        line::{StyledLineIterator, UniformSpaceConfig},
+        cursor::Cursor, line::StyledLinePixelIterator, space_config::UniformSpaceConfig,
         StateFactory, StyledTextBoxIterator,
     },
     style::height_mode::HeightMode,
@@ -34,7 +33,7 @@ where
     NextLine(Option<Token<'a>>, Cursor<F>, Parser<'a>),
 
     /// Renders the processed line.
-    DrawLine(StyledLineIterator<'a, C, F, UniformSpaceConfig, CenterAligned>),
+    DrawLine(StyledLinePixelIterator<'a, C, F, UniformSpaceConfig, CenterAligned>),
 }
 
 impl<'a, C, F, V, H> StateFactory<'a, F> for StyledTextBox<'a, C, F, CenterAligned, V, H>
@@ -78,7 +77,7 @@ where
                             .measure_line(parser, carried_token.clone(), max_line_width);
                     cursor.advance_unchecked((max_line_width - width + 1) / 2);
 
-                    self.state = State::DrawLine(StyledLineIterator::new(
+                    self.state = State::DrawLine(StyledLinePixelIterator::new(
                         parser_clone,
                         cursor,
                         UniformSpaceConfig {
@@ -97,7 +96,7 @@ where
                     self.state = State::NextLine(
                         line_iterator.remaining_token(),
                         line_iterator.cursor,
-                        line_iterator.parser.clone(),
+                        line_iterator.parser(),
                     );
                 }
             }
